backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.2",
    "openai>=2.30.0",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.42.0",
]
